TFG_AWS_ANTIVIRUS/infraestructura-AWS/AWS-Services/EC2-Ubuntu/docker-infraestructure/ConTrol/requirements/requirements.txt
mysql-connector-python~=9.4.0
cryptography~=46.0.2
uvicorn